Improve automated relevancy judgement generation
This makes an improvement to how we generate training
data from Google Analytics click data.

Rather than using heavily sampled Google Analytics
aggregate click through rate data for each query, which
leads to inaccurate data for training on, this uses
a CSV of BigQuery data and generates a training dataset
from this.

The training dataset computes the estimated CTR at position 1
'CTR@1' for each query:document pair. CTR@1 is the percentage
of clicks a link would get if we put it at the top of the results.

The CTR@1 is a hack, to get around using a click model which
might be more accurate.

We turn the CTR@1 for each result into a score between 0-3,
a relevancy score which is equivalent to the manual relevancy
scores that we have collected previously.

Finally the relevancy judgements have features embedded and
are turned into LibSVM formatted data, for training the model.

https://trello.com/c/VuvDR3Ac/1143-automatically-generate-relevancy-judgements

Later we can automate the fetching of BigQuery data, to make
it possible to run a single command to generate a model trained
on the latest click data.