Remove unconfirmed implementation detail from LTR ADR
I removed the implementation details that haven't been confirmed
as these can be the subject of further ADRs, and we can reach
decisions on implmentation when we are further along with the
project and have validated that it will be successful.

Since these decisions around implementation haven't been made
yet I don't think these parts add value to the ADR, as it will
become out of date and may confuse devs in the future.