Don't send viewcounts and rank_eval for individual queries to graphite
This is taking up a fair amount of disk space on graphite.
For example viewcounts took up 4GB.

We aren't using viewcounts, and recording rank_eval for
individual queries isn't that useful right now.

rank_eval might be useful in future, because we'd be able
to see how relevancy for a query had changed over time.

https://trello.com/c/3e0NSK06/1162-stop-logging-view-count-and-rank-eval-for-individual-queries