Introduce autocomplete functionality to Search-API
This pull request includes base debug functionality to enable
autocomplete suggestions in Search-API returns using the
"debug=autocomplete" flag. This is beta functionality awaiting
further tuning, exemptions and discussion.

This is done via a tuned edge-ngram method based on the below:
-https://www.elastic.co/guide/en/elasticsearch/reference/master/analysis-edgengram-tokenizer.html
-https://medium.com/@taranjeet/elasticsearch-building-autocomplete-functionality-494fcf81a7cf

Once merged into master, search will have to be reindexed to allow
the functionality to be enabled. Without a reindex, no results will
be returned.