Enable LTR using string env variables
We were previously relying on the truthiness of `ENV["ENABLE_LTR"]` to control
enabling learn to rank.  We were trying to set the env var to a boolean in
puppet, but that's not possible.

We'll do a check for the string "true" instead (to avoid an env var of "false"
accidentally passing the truthiness rules) and convert puppet to setting
strings instead.