Send NDCG@1,3,5,10,20 to graphite
This adds a new rake task that will run on a schedule:
relevancy:send_metrics_to_graphite. It will send NDCG
to graphite, to populate a new metric.

This is necessary since it uses the new Evaluate::NDCG
class that computes NDCG manually (rather than using
the rank_eval endpoint provided by Elasticsearch), and
that we can use to compare the Learning To Rank model
to our existing search relevancy performance.