Move stats radio button facet to the top
At present, it looks like the document type field is under the "Topic" heading
in the filters section.  This fixes that, and also makes it nearer the top
on mobile.  It's more used than Topic on the stats finder, so this makes sense.

We'd originally kept topic as the top filter to keep things broadly consistent
between supergroup finders as we expected to build some navigation to allow
users to flip between them.  This hasn't happened, so let's do a small
thing to make life easier for people.