Add a module to run the full data pipeline process
Fetches analytics data, turns it into training data, and uploads it to
S3.  Uses redis to ensure that only one job runs at a time, as this
isn't terribly fast.

The intent is to have API endpoints to trigger a job and to check job
status, which Concourse will poll to wait for this part of the
training / deployment pipeline to complete.