Use lazy enumerators in LTR training data generation
The current training data generation uses a lot of memory, as it
holids in memory all at the same time:

- Bigquery results
- Relevancy judgements
- Augmented relevancy judgements
- SVM data

And there's other things (like SVM data in another form) which
temporarily live in memory, before being garbage collected.  That's a
lot to hold in memory at once.

This commit changes the data generation to use lazy enumerators, so
everything other than the bigquery data (which is still held in memory
in full) is only computed when needed.

Unfortunately, we can no longer print out a progress metric when
computing the augmented relevancy judgements, because to get the total
number of judgements would require forcing that iterator, which
defeats the point.

Finally, I've reduced the number of queries cached in the
judgement-augmentor to 25, as we're getting quite a lot of data from
elasticsearch at a time.