Allow expansion of role facets
This adds a registry for ministerial roles (as they are the only kinds
of roles you can tag documents to) and adds it to the entity expanders.

This means that when seeing results for a document tagged to a role, we
can see the expanded fields for it.

For example at the moment you can only see the `slug`
(https://www.gov.uk/api/search.json?facet_roles=1500,examples:0,order:value.title&count=0):

```json
{
  "value": {
    "slug": "attorney-general"
  },
  "documents": 1
},
```