Introduce smarter autocomplete results through popularity weighting
Unlike the actual search functionality, Elasticsearch suggest API does
not allow automatic results ranking on any "suggest" fields used in
the Elasticsearch schema for autocomplete functionality. This combined
with a need for speed resulted in a "dumb" original implementation.

This pull request improves the previous implementation in the following
key ways:

- Allows us to use the suggestion API as provided by Elasticsearch
- Allows us to use the popularity of a given document to weigh
  autocomplete results
- Allows us an easy means to keep popularity of autocomplete results
  concurrent with the base popularity used for ranking.
- Opens the door for potential improvements in using context-based
  suggestion ranking.

The autocomplete field is updated when the update_popularity script is
called, ensuring concurrent values in both fields.