Deployment stability improvements
This makes a couple of changes to the unicorn config in
an attempt to improve deployment stability.

In order to increase application startup times, this adds
https://github.com/Shopify/bootsnap

Boot times are improved by using bootsnap, but initial requests
to a newly booted worker are still slow;
this change attempts to mitigate the impact of this.

https://trello.com/c/CVrZaaPv/1320