Enable deploying models using tag name
Rather than deploying the latest trained model (provided by the
train job), this enables the deploy script to use a different
model stored in S3 by looking at the MODEL_TAG envvar.
It then creates an endpoint in the environment using the same
tag, so the endpoint will be e.g.
integration-search-ltr-endpoint-variant-b
that will need to be added to puppet.