Add Healthcheck for Elasticsearch node drive space
Typically we only discover that Elasticsearch has run out of diskspace
when we start to see errors. This usually happens because there are old
Elasticsearch indices that have not been removed with the relevant
rake task clogging up the system.

This healthcheck will test the overall space available and cause an
alert if we have less than 20% free on the disks. A percentage decided
upon based on current usage trends.