Add deploy b-variant model job to concourse
This enables us to deploy a named model to Sagemaker.

It looks up the model using the model_tag envvar
in s3 and deploys it to a separate endpoint in
Sagemaker.

The idea is that we'd create a new model ready for
testing (e.g. a new feature), so we'd create a new
model and put it in s3. Once done, we could create
a concourse job to deploy the model. Finally, we
could deprecate the original endpoint and use only
this model. We may want to use a 'latest' tag on
the current endpoint.

https://trello.com/c/1EjO05Dj/1333