Enable sagemaker endpoint selection in queries
Now users of search api can set ab_tests=mv:B, which will
be used by search-api to determine which model should handle
the request.

This requires enabling an endpoint in the form
'<default endpoint>-<variant name>'. In the above instance
one would need to create a model called B and deploy it to
sagemaker, and then permit the use of this endpoint in
Search API.

https://trello.com/c/1EjO05Dj/1333