Automate deletion of old unused indexes
We were having issues wherein the reliance on manual deletion of old
indexes from the Elasticsearch system was clogging up the storage space
of our cluster. This commit adds a Rake task designed to mitigate this
problem.

The task is designed to be run on schedule in Jenkins. It relies on the
ability to source the "updated_at" value from the last edited record on
a given index. With this in mind, we have two steps.

1. We gather all indices under the given index name, ignoring the last
two that were created, into a map. The last two that were created are
almost universally guaranteed to be the currently used index, and
the index it replaced.

2. Based on whether we have access to an "updated_at" field on the given
index, we either:

- Check if the age of the index as given by the most recent "updated_at"
 value is older than N days as given by the environment variable
 MAX_INDEX_AGE and if it is, delete it. Current indexes: government,
 govuk, detailed

- If we do not have access to the field, we simply remove the extra
wasted indexes as we have no way to accurately gauge the age of them.
Current indexes: metasearch, page-traffic

This should effectively manage overall storage use on our cluster, once
the relevant Jenkins job is created.