Do not default to tensorflow serving
Previously, when sagemaker isn't configured, search-api would assume
it can talk to a tensorflow serving container, on 0.0.0.0 if all else
fails.  This commit makes it instead default to not reranking at all
if neither sagemaker nor serving have their env vars set.

The serving stuff is remaining, because it's useful for local
development and it's only a small amount of code, but we're not using
it in production any more.