Restructure README
This moves some content around in the README.

I've removed the details about working with
Search API within the VM as this is no longer
supported.

I've added stub documents for explaining
how indexing and searching works, which should
be completed.

We also still need to add details about Learning
to Rank to the searching and relevancy documents,
in addition to a new document about the learning
to rank infrastructure.