Remove data pipeline worker
The worker tried to do everything in a lazy fashion, which is good for
keeping memory usage down, but bad for saving intermediate results.

Memory usage isn't a concern if we're running this process outside of
a live search-api instance, so there's no need for a clever worker.
Instead, Concourse will run each rake task in sequence, and save all
the results (final and intermediate) to S3.