Combine results at the same rank for relevancy scoring
A set of results will only appear at the same rank for the same query
if 1 of them is a page and the rest of them are parts on that
page (eg, a guide).

Parts aren't separate entities in elasticsearch, so their relevancy
needs attributing to the parent page.

First, work out the URL of the parent page (this is easy because part
URLs are always of the form "{parent url}/{slug}"), and sum all the
clicks.  For the views, take the max.