Use HTML attachments as parts, if parts are not given (whitehall)
A bunch of the coronavirus guidance is being published as attachments
of generically-named guidance pages.  That's good if you know what
you're looking for, but can be confusing if you're not.

So we want to display HTMl attachments as parts, but only if the
result doesn't have any parts already.