Don't try to turn attachments with missing URLs into parts
Until the whitehall part of this PR is deployed, whitehall will
continue to send attachments that have content but no URL.  So we need
to be resilient to that.

One option would be to just deploy the changes in quick succession and
reindex any documents published in the gap, but being resilient to
unexpected input is good practice anyway.