Fix linting issue with top-level methods
We should avoid top-level methods, as they break code encapsulation.
This replaces the use of top-level methods in several rake tasks by
replacing the call to the method with the code of the method. While
this makes a couple of tasks more complicated, we can always refactor
these in future if they need maintenance.