Only turn attachments into parts if the URL matches the parent
We don't index part URLs, only slugs, and assume that all parts have a
URL of the form `{parent_url}/{slug}`.

This is usually correct, but not for HTML attachments on statistics
pages.  There the parent page is under /government/statistics/..., but
the attachment is under /government/publications/....  We could fix
that by storing URLs, but that'd require reindexing a lot of stuff, so
for now let's just go for the simpler solution of rejecting those
attachments.