Add parts to the sitemap
This checks to see if a document has parts in its search entry. These represent
different pages within the site, such as chapters in a guide or HTML attachments
to a publication.

We're adding them here so that they can have a priority assigned to them.

They are less important than their parent document, but also need to inherit
any boosts that the parent gets, either for formatting, withdrawn status,
or anything else.