Process document and each of its parts similarly
This assembles all the parts associated with a document, and the "base" document.
We then use the same process for each to add to the latest sitemap "chunk" and
subsequently check for whether we've reached our maximum chunk size.

Each chunk is used to make a separate sitemap.xml file which may hold a maximum
of 25_000 items (at the time of writing).

The [spec allows for up to 50_000 entries per file](https://www.sitemaps.org/protocol.html),
but opening an XML file of that size is not very friendly, so I can understand
why we've previously chosen a smaller limit.