Use 6 months of data to train model
Pre-December 2019 we used a different field for search links in
Google Analytics (not product.v2ProductName AS link).

The current code for fetching searches from BigQuery does not
account for this, so fetching query document pairs from pre-2020
using the current query will return "(not set)" as the document
link. The query should be amended so that 2019 data can be used.

This is a quick fix to ensure that we only use the newer field
for generating training data for the model.