Revert "Combine results at the same rank for relevancy scoring"
This reverts commit c3f14e86205f67895a7810faa052559f2b08ac81.

I'm investigating why Search CTR and NDCG declined in March 2020.
This coincided with changes in user behaviour in response to COVID-19,
so changes may be required to support the new behaviour. However, as a
first pass I'm investigating whether any releases correlated with the
decrease in our offline and online metrics, starting with PRs #2041 and #2044.

We found that reverting commit c3f14e8 (PR #2041) improves NDCG@1,10,20
by 1-2% and has a neutral effect on NDCG@3,5.

As we don't have a search team at the moment we're not running proper
AB tests so will monitor the impact of this on CTR on 100% of traffic.