Use live links when preparing tags
Otherwise the Publishing API returns the expanded links for the draft
if one exists, which may be different from the live content. The
Search API is currently just deployed to index live content, so this
should fix issues where Whitehall content is indexed with the
organisations for the draft editions.