Make LTR Sagemaker instance types larger
We could do with a little more headroom in production. Moving
to a C5 type removes the CPU credit quota. LTR is more compute
heavy than anything else.

We'll monitor for a bit, and move back down to C5.large if it
seems overkill.

I've changed staging too so that I can test the deploy.