Fix ClassVars issue
This hook appears to be redundant since it was added in 6bf9d42a9:
I can run the tests without issues. We can always revert and look
at other fixes if this spec does become flakey.

Note that, while investigating the effect of this hook, I noticed
that it's unlikely to have any: the "registries" variable is only
used to instantiate two other **memoised** objects.