Adjust sagemaker instance type and count
At present, the sagemaker instance type will always be whatever was set
originally. This changes things so that the instance type will be updated
if the configuration changes. The current live machine type is a t2.medium.
Changing production from c5.xlarge to c5.large so that we can go up in increments
and save a bit of cash.

I want to test how this works in integration/staging before moving to production, as
it would need to handle in-flight upgrades gracefully, or we'll have to
do some more work.

Learning to rank is currently disabled in Puppet, so I'll need to turn that
on first to do testing, but that does mean we can deploy this to production
without worrying too much.