Remove task_list document type
This was what step by step pages were originally called. There are no instances
of this type anywhere on GOV.UK. It was [removed from govuk-content-schemas](https://github.com/alphagov/govuk-content-schemas/commit/46d5d234029a9f857f37876005d477982297f8ca)
but sadly, that undoing didn't get as far as here.

I found this by looking at the Elasticsearch slow query logs and noticed that
it was included in the list of formats that we filter for.