Remove need for REDIS_HOST and REDIS_PORT env vars
This is one of the few GOV.UK projects that uses this old approach to
configuring Redis - whereas most of the other ones use the approach
supported natively by the Redis gem [1] of using REDIS_URL. Removing
this will allow us to remove the need for REDIS_HOST and REDIS_PORT env
vars in our architecture and also allows this code to be simplified.

This is a second attempt at this following [2], that attempt was
reverted because it broke the end-to-end tests. I've since learnt the
problem there was each Search API process in e2e-tests is set with a
different GOVUK_APP_NAME [3]. We were previously configuring Sidekiq
with that env var, and having that differ meant that the app and worker
were listening to different queues.

A difference to the previous iteration is also adding an explicit
REDIS_URL, this seems to make this code easier to understand without
needing to modify govuk_sidekiq further.

[1]: https://github.com/redis/redis-rb
[2]: https://github.com/alphagov/search-api/commit/8fe57f8dbade30a56a6276ccaa68ac0ce97b185b
[3]: https://github.com/alphagov/publishing-e2e-tests/blob/7d5ce8602c6bd6111606ee6a3f17bacfbbc4cb01/docker-compose.yml#L87-L177