Hide the show only Brexit results checkbox
Affects: https://www.gov.uk/search/news-and-communications

We'll follow this up in finder-frontend to formally retire this feature.

We're retiring it because it has a fair bit of special handling which adds
complexity to search. This is a risk when there is no specialist team working
in this area, so we've chosen to retire it instead, as not many people are
using it.