Remove the pins from Gemfile
As per the GOV.UK convention [1] we shouldn't use pinning unless it has
a specific reason which is documented. The only gem that seems to meet
this criteria here is Elasticsearch which needs to be pinned to 6.x
releases as only they work with Elasticsearch server version 6, which we
use.

[1]: https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#gemfile-organisation