Move rainbow to main gems
This gem is required in the debug rake task file which is loaded
whatever environment you are in. Therefore to run any rake tasks you
should have this gem.

This seems to be causing a problem with publishing-e2e-tests that tries
to run a rake task and this is failing. I've no idea how this used to
work in the past, but this seems like how this should be configured.