Remove deprecated supertypes
We don't use these supertypes any more. Content purpose supertypes are used
all over the site and are maintained whereas these are starting to get out
of date.

I've updated the boosting.yml file using the more up to date supertypes
and removed references to the older types so we'll no longer respond to
queries for them.