Replace deprecated supertypes in boosting
navigation_document_supertype has not been updated much in recent years
and several new content types have arrived since. It's not really used
around the system so it makes more sense to replace it with
content_purpose_supertype references.

I've updated the spec too. Guidance content is still boosted, but I think
this spec was more about asserting that _something_ was boosted correctly
rather than worrying about exactly what it was. With that in mind, I've
switched the spec to checking that services get a bump.