Dedupe boosting further
In https://github.com/alphagov/search-api/pull/2284 I removed a couple of
out of date supertype groups with more current ones, however it wasn't quite
equivalent.

In this commit, I remove the content_purpose_subgroups I added, and replace
them with direct additions of the content types that were not already present
in the content_purpose_supergroup services group, or elsewhere tracked
in this file.

I've dropped the boosting of `contact` entirely because it was boosted by
2.5 and also suppressed by 0.3.

I've removed step_by_step_nav and smart_answers as direct references because they're
picked up in the boosting of content_purpose_supergroup services group.

I've left transaction as being boosted, as it always was prioritised over
other content.

document_collection was duplicated, so I've given it a lift, and removed the
duplication.

The other additions are direct copies from the original configuration.

I've chosen to downgrade to a boost of 2 in general for these, because we
want to gradually test removal of boosting, as it's unscientific.