Fix broken docs link and consistently link within repo
This fixes the link which had changed from
https://docs.publishing.service.gov.uk/apps/search-api/using-the-search-api.html
to https://docs.publishing.service.gov.uk/repos/search-api/using-the-search-api.html.
Rather than continuing to link to the developer docs it seemed more consistent
and less fragile to link to doc files within this repo (as per the various
other documentation links in this readme).