Create DRCF digital markets research finder
Following CMA request to create a specialist finder to bring together research
from several organisations.

This introduces search filters that are prefixed with
digital_market_research.

We expect digital_market_research_category to be a choice of two fields (thus
singular) and digital_market_research_publish_date to be a date field.
We expect the other fields to be arrays of strings.

Co-authored-by: Barbara Slawinska <barbara.slawinska@digital.cabinet-office.gov.uk>