Update mapped document type for product_safety_alert_report_recalls
This document [1] is present in publishing API and in the content store, but it has not been
added to our Elasticsearch index, ie:
/api/search.json?filter_link=/product-safety-alerts-reports-recalls/product-safety-report-the-mini-mermaids-dolls-toy-2101-0112
returns nothing.

This is because of a typo in the mapped_document_types yaml. This file is used by SearchAPI to construct our ES queries, it maps
the publishing_api document type to the ES concept of document_type. In this case it's mapping requests for all the docs we've
published with a doc type of product_safety_alert to docs in the index with a doc type of product_safety_alert_report_recall.

Product_safety_alerts was the original publishing_api document type, but this was renamed here [2]

Because we are using an existing ES schema, I don't think we will need to reindex search [3] following this PR's merge.
I will need to republish all of the documents of this type [4], but there aren't many.

[1] /api/content/product-safety-alerts-reports-recalls/product-safety-report-the-mini-mermaids-dolls-toy-2101-0112
[2] https://github.com/alphagov/specialist-publisher/pull/1950/files
[3] https://docs.publishing.service.gov.uk/manual/make-a-new-document-type-available-to-search.html#3-reindex
[4] https://docs.publishing.service.gov.uk/manual/make-a-new-document-type-available-to-search.html#4-republish-all-the-documents