Use env `LOG_TO_STDOUT` to enable stdout logging
In EKS, we want logging to go to stdout rather than to a logging
file. This commit introduce a new `LOG_TO_STDOUT` env which
when set will log to stdout.

By default, stdout logging is enable in the Docker image as decided
[here](https://github.com/alphagov/govuk-helm-charts/pull/23).
We need to disable that for publishing-e2e.