Remove publishing e2e tests from Jenkinsfile
This commit removes the E2E tests from this app's CI pipeline. Once merged,
the Jenkins CI server will no longer kick off a run of the
publishing-e2e-tests job when new PRs are opened.

This change is the first step in the overall decommissioning process,
and will be applied to every application currently running the E2E tests.