Add ability to boost custom fields
We've initially decided to boost custom fields above indexable_content
(body etc) but below description for the time being.

We may want to increase the boost value if we feel it is appropriate.

`field_boost` param will include the more important fields (if any)
which will feed into the boosting query.

We could implement dynamic boosting (`boost_field` specifies field AND
boost value) but this would mean boost config existing in multiple
places (requesting app and here) which would be more of a pain to
maintain and possibly have unintended consequences if the values change
in the future.

This change applies to quoted and unquoted queries for consistency.