[1141] - Add Marine Equipment Approved Recommendation as a new specialist document
Added necessary fields and schemas to search for Marine Equipment recommendations
The fields Year Adopted, Keyword, and Reference Number are free text fields which are searchable text.
Original requirements asked for them to be filterable.
However, due to Elasticsearch having to be able to reindex over new free text values, and they are non aggregated fields, meaning filtering on free text is currently not supported functionality.
There is an existing search box to free text search to serve lookup requirements.