Upload ltr datasets to S3
This extends the learn to rank tasks to upload generated dataset to S3,
instead of seperately in the jenkins script. This removes the dependency
of aws-cli to be installed on the system and reuses the aws gem
dependency.

Files will only be uploaded to S3 if the "EXPORT_TO_S3" env var is set.