Prepend a Time class extension
This change fixes a breaking change introduced in v3.11.4 of the oj gem.

There is quite a gap between v3.11.3 and v3.15.0, so the first step was to apply a binary chop on the versions between v3.11.3 and v3.15.0 to identify where the break originates.

The break was introduced in 3.11.4 - here’s the [GitHub compare](https://github.com/ohler55/oj/compare/v3.11.3...v3.11.4).

Having found the breaking change, the next stage was to look at the errors.

When we run the tests we will see multiple failures of the following kind...

```bash
Failures:

  Failure/Error:
    client(cluster:).get(
      index:,
      type:,
      id:,
    )
    Elasticsearch::Transport::Transport::Errors::NotFound: [404]

     # ./spec/support/integration_spec_helper.rb:254:in `fetch_document_from_rummager'

  1) ChangeNotificationProcessorTest triggering a reindex
      {"_index":"government_test-2023-06-06t08-05-16z-8b7d295d-58aa-4808-bd6b-6fcaa54f350f","_type":"_all","_id":"/foo","found":false}
      # ./spec/integration/indexer/change_notification_processor_spec.rb:25
      ...
```

All of these errors point to the same line in the code `./spec/support/integration_spec_helper.rb:254`, but are otherwise not terribly helpful.

If you Google this error [there are suggestions](https://stackoverflow.com/questions/49810704/elasticsearchtransporttransporterrorsnotfound-404-errorroot-cau) that `Model.import(force: true)` will resolve the problem - it does not! Nor is it an ElasticSearch problem, despite the wording of the error message.

Now, purely on a hunch we re-tried v3.15.0 just to see if the errors differ, get worse etc from v3.11.4. Interestingly, they appear to be the same. This is good news - if we can solve v3.11.4, then v3.15.0 should also be fixed.

It was around this time when another whim made us look at the elasticsearch container log in Docker Desktop while running the tests. We noticed this message for multiple documents - and they appeared to correspond to failures in the tests...

```bash
Caused by: java.lang.IllegalArgumentException: Invalid format: "2023-06-09 09:44:19 +0000" is malformed at " 09:44:19 +0000"
at org.joda.time.format.DateTimeParserBucket.doParseMillis(DateTimeParserBucket.java:187) ~[joda-time-2.10.1.jar:2.10.1]
[DEBUG][o.e.a.b.TransportShardBulkAction] [lYwGut9] [government_test-2023-05-22t15-10-16z-afdcb65c-d945-43f2-a333-ce01eab5fc77][2] failed to execute bulk item (index) index {[government_test][generic-document][/something-not-in-publishing-api], source[{"link":"/something-not-in-publishing-api","updated_at":"2023-05-22 15:11:12 +0000","document_type":"edition","popularity":0,"popularity_b":1,"view_count":1,"format":"edition","user_journey_document_supertype":"thing","email_document_supertype":"other","government_document_supertype":"other","content_purpose_subgroup":"other","content_purpose_supergroup":"other"}]}
org.elasticsearch.index.mapper.MapperParsingException: failed to parse field [updated_at] of type [date] in document with id '/something-not-in-publishing-api'
```

We proved the hypothesis that the ruby test console error and the elasticsearch log error are related, by simply running a single failing test multiple times. Sure enough, we got the same test and elasticsearch error each time.

Now that we know they are related, what is the problem?

Elasticsearch is written in Java and is uses a package called [`joda-time`](https://www.joda.org/joda-time/) - this is where the error originates. That error is then bubbled-up to ruby test output. `joda-time` is looking for a particular datetime format and cannot parse the datetime format it's is now being given by v3.11.4.

Back to the oj gem and we see [this commit](https://github.com/ohler55/oj/commit/9d02e8a8d3c025839abcfcc828c2dd35b37a3036) monkey patches the Date class in `lib/oj/mimic.rb`.

So, how did the date time format change...

We added the following in a test and re-ran that single test...

```ruby
puts Time.current.as_json
puts Time.current.to_json
puts JSON.generate(Time.current)
```

...and got...

```bash
2023-06-09T07:55:08.141+00:00
"2023-06-09T07:55:08.141+00:00"
"2023-06-09T07:55:08.141+00:00"

2023-06-09T08:13:17.347+00:00
"2023-06-09T08:13:17.348+00:00"
"2023-06-09T08:13:17+00:00"  # WRONG!
```

From this we knew that `JSON.generate(Time.current)` is in the wrong format. However, we can ignore the `JSON.generate()` part - as that just JSONifies whatever is passed in, so the error is really in `Time.current`.

What is passed in, is a string representation of the current time via a `Time` object. The string representation is obtained by the standard `to_s` method. So, we need to apply the desired format (`strftime("%Y-%m-%dT%H:%M:%S.%L%:z")`) to all instances of the `Time` class.

Yes, we need to monkey patch the `Time` class 🙈
Yes, we know about monkey patching and why you shouldn't do it 👍
No, we have no choice 🤮

So, let's do this in the most humane way possible.

Ruby has a `prepend` [method](https://ruby-doc.org/3.2.2/Module.html#method-i-prepend) that "Invokes `Module.prepend_features` on each parameter in reverse order" - a fancy way of saying you can change methods on a class - which was added to support changing classes in a graceful manner. This looks like the least invasive approach and is considered by some to the the preferred method of monkey patching.

In our defence, we would point out that the reason v3.11.3 worked is because it contained and monkey patch. The reason v3.11.4 fails is because that monkey patch was removed. Elasticsearch (actually `joda-time`) needs datetime's in a specific format that we cannot change. So essentially, we are just re-applying the previous monkey patch - albeit in a different way.

![Alt Text](https://media.tenor.com/Fi_oI34mE8MAAAAM/hansolo-notmyfault.gif)

* 2 issues [658](https://github.com/ohler55/oj/issues/658) and [701](https://github.com/ohler55/oj/issues/701) where created against v3.11.4 of the oj gem.
* The oj gem was originally added for performance reasons.
* JSON parsing and creation has improved immeasurably since then, so the oj gem may no longer be required. We have chosen to leave it in place - for now - for the same reasons.
* The oj gem can be removed completely, but it still requires the `Time` extension added here. This may be preferable - if performance concerns are proved to no longer be an issue.
* We no longer need to pin the version of the oj gem. This has been removed.

[Trello](https://trello.com/c/l7X4Yt8C/455-fix-search-api-oj-dependabot-issue)